Adds preview artwork to upload and edit artwork pages

+++ client/src/PreviewArtwork.js
@@ -2,45 +2,47 @@ import React, {useState} from 'react'
 import {useHistory} from 'react-router-dom'
 import './PreviewArtwork.css'
 
-function PreviewArtwork({id, image_url, tags, likes, title, user, likedByUser, updateUserLikedArtworks}) {
+function PreviewArtwork({id, image_url, tags, likes, title, user, likedByUser, updateUserLikedArtworks, preview}) {
   const [hovered, setHovered] = useState(false)
   const [updatedLikes, setUpdatedLikes] = useState(likes)
   const history = useHistory()
   
   function handleLike() {
-    const artwork = {
-      id: id,
-      likes: updatedLikes,
-      title: title,
-      image_url: image_url,
-      user: user
-    }
-    if(likedByUser === false) {
-      fetch("/user_likes", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ artwork_id: id })
+    if (!preview) {
+      const artwork = {
+        id: id,
+        likes: updatedLikes,
+        title: title,
+        image_url: image_url,
+        user: user
+      }
+      if(likedByUser === false) {
+        fetch("/user_likes", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ artwork_id: id })
+        }).then((r) => {
+          if (r.ok) {
+            r.json().then(() => {
+              
+              setUpdatedLikes(() => updatedLikes + 1)
+              updateUserLikedArtworks(artwork)
+            })
+          }
+        })
+      }
+      else if(likedByUser === true){
+        fetch(`/user_likes/${id}`, {
+          method: "DELETE"
       }).then((r) => {
-        if (r.ok) {
-          r.json().then(() => {
-            
-            setUpdatedLikes(() => updatedLikes + 1)
-            updateUserLikedArtworks(artwork)
-          })
+        if(r.ok) {
+          setUpdatedLikes(() => updatedLikes -1)
+          updateUserLikedArtworks(artwork)
         }
       })
-    }
-    else if(likedByUser === true){
-      fetch(`/user_likes/${id}`, {
-        method: "DELETE"
-    }).then((r) => {
-      if(r.ok) {
-        setUpdatedLikes(() => updatedLikes -1)
-        updateUserLikedArtworks(artwork)
       }
-    })
     }
   }
   return (
@@ -60,10 +62,18 @@ function PreviewArtwork({id, image_url, tags, likes, title, user, likedByUser, u
             ?
             <div>
             <h3
-              onClick = {() => history.push(`/users/${user.username}/artworks/${id}`)}
+              onClick = {() => {
+                if (!preview) {
+                  history.push(`/users/${user.username}/artworks/${id}`)
+                }
+              }}
             >{title}</h3>
             <p
-              onClick = {() => history.push(`/users/${user.username}`)}
+              onClick = {() => {
+                if (!preview) {
+                  history.push(`/users/${user.username}`)
+                }
+              }}
             >{user.username}</p> 
           </div>
           : null
